Move profile stats into a separate component
